add data quality test in vectorizer loop and address parser
