+++ deepparse/tools.py
@@ -1,8 +1,12 @@
 import os
 import warnings
+from typing import List
 
 import poutyne
 import requests
+
+from .data_error import DataError
+from .data_validation import is_whitespace_only_address
 
 BASE_URL = "https://graal.ift.ulaval.ca/public/deepparse/{}.{}"
 CACHE_PATH = os.path.join(os.path.expanduser("~"), ".cache", "deepparse")
@@ -106,3 +110,29 @@ def handle_model_path(checkpoint: str) -> str:
         )
 
     return checkpoint
+
+
+def validate_data_to_parse(addresses_to_parse: List[str]) -> None:
+    """
+    Validation tests on the addresses to parse to respect the following two criteria:
+        - no addresses are empty strings, and
+        - no addresses are whitespace-only strings.
+    """
+    if empty_address(addresses_to_parse):
+        raise DataError("Some addresses only include whitespace thus cannot be parsed.")
+    if whitespace_only_addresses(addresses_to_parse):
+        raise DataError("Some addresses only include whitespace thus cannot be parsed.")
+
+
+def empty_address(addresses_to_parse: List[str]) -> bool:
+    """
+    Return true if one of the addresses is an empty string.
+    """
+    return "" in addresses_to_parse
+
+
+def whitespace_only_addresses(addresses_to_parse: List[str]) -> bool:
+    """
+    Return true if one the address is composed of only whitespace.
+    """
+    return any([is_whitespace_only_address(data) for data in addresses_to_parse])
