+++ deepparse/vectorizer/train_vectorizer.py
@@ -21,10 +21,12 @@ class TrainVectorizer:
         """
         input_sequence = []
         target_sequence = []
+
         input_sequence.extend(
             self.embedding_vectorizer([address[0] for address in addresses])
-        )  # need to be pass in batch
-        # otherwise the padding for byte-pair encoding will be broken
+        )  # Need to be pass in batch
+
+        # Otherwise, the padding for byte-pair encoding will be broken
         for address in addresses:
             target_tmp = [self.tags_vectorizer(target) for target in address[1]]
             target_tmp.append(self.tags_vectorizer("EOS"))  # to append the End Of Sequence token
