+++ tests/vectorizer/test_train_vectorizer.py
@@ -1,7 +1,9 @@
 from unittest import TestCase
 from unittest.mock import MagicMock
 
-from deepparse.vectorizer import TrainVectorizer
+from deepparse.data_error import DataError
+from deepparse.embeddings_models import EmbeddingsModel
+from deepparse.vectorizer import TrainVectorizer, BPEmbVectorizer
 
 
 class TrainVectorizerTest(TestCase):
@@ -17,3 +19,13 @@ class TrainVectorizerTest(TestCase):
 
         output = train_vectorizer(["A list of"])
         self.assertEqual(list(output), [(0, [0, 0])])
+
+    def test_givenAVectorizer_whenCallWithAnWhiteSpaceOnlyAddress_thenRaiseError(self):
+        embedding_network = MagicMock(spec=EmbeddingsModel)
+        embedding_network.dim = 2
+        bpemb_vectorizer = BPEmbVectorizer(embedding_network)
+
+        train_vectorizer = TrainVectorizer(bpemb_vectorizer, MagicMock())
+        a_whitespace_only_address = " "
+        with self.assertRaises(DataError):
+            train_vectorizer([a_whitespace_only_address])
